refactor: split IR builder helpers
